pyproject.toml
[project]
name = "langchain-react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2026.1.4",
    "faiss-cpu>=1.13.2",
    "langchain>=0.3.0",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.16",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
]
